==> engenharia-dados/formacao/engenharia-dados-python-banco-dados/Google BigQuery/01.conexao_com_autentificador_fisico.ipynb <==
#pip install  google-cloud-bigquery
#pip install --upgrade google-cloud-bigquery
#pip install  pandas_gbq
#pip install  pandas

from google.cloud import bigquery
import pandas as pd

# Especifique o caminho para o arquivo JSON de credenciais
credencial = "/Users/juniorsilva/Downloads/python-curso-459117-fc7b944c47c6.json"

# Crie um cliente BigQuery usando as credenciais
client = bigquery.Client.from_service_account_json(credencial)

# Construa sua consulta
query = """ 

    SELECT
    id,
    first_name,
    last_name,
    email
    FROM `python-curso-459117.AulaPython.Clientes`
    WHERE state IN ('Piauí', 'Alagoas', 'Amazonas')

"""

# Execute a consulta
resultado = client.query(query)

"""
for row in resultado:
    print(row)
"""

# Converta os resultados em um DataFrame
df = resultado.to_dataframe()



df.head(5)
