==> engenharia-dados/formacao/engenharia-dados-python-banco-dados/Google BigQuery/04.carregando_dados_para_google_bigquery.ipynb <==
#pip install  google-cloud-bigquery
#pip install --upgrade google-cloud-bigquery
#pip install  pandas_gbq
#pip install  pandas

#pip install pandas pandas_gbq

import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq

# Carregue suas credenciais
credencial = service_account.Credentials.from_service_account_file(
    "/Users/juniorsilva/Downloads/python-curso-459117-fc7b944c47c6.json", scopes=['https://www.googleapis.com/auth/bigquery'])

df = pd.read_excel("/Users/juniorsilva/develop/data/estudos-engenharia-dados/python-etl/Arquivos Excel/arquivos_excel/Produto.xlsx")
df.head(5)

df.to_gbq(destination_table='python-curso-459117.AulaPython.Produto', # caminho completo do dataset + nome da tabela
          project_id='python-curso-459117',
          if_exists='replace',
          credentials=credencial)

# replace = substituir a antiga base de dados pela nova
# append = adiciona dados no final da tabela existente
