==> engenharia-dados/formacao/engenharia-dados-python-banco-dados/Google BigQuery/07.carregando_varios_arquivos_e_transformando.ipynb <==
import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq

# Carregue suas credenciais
credencial = service_account.Credentials.from_service_account_file(
    "/Users/juniorsilva/Downloads/python-curso-459117-fc7b944c47c6.json", scopes=['https://www.googleapis.com/auth/bigquery'])

# Nome do projeto e dataset
projeto = 'python-curso-459117'
dataset = 'Ecommerce'

# Mapear arquivos com nome da tabela
arquivos = {
    'produto': '/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Produto.xlsx',
    'ordens': '/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Ordens.xlsx',
    'items': '/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/items.xlsx',
    'categoria': '/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Categoria.xlsx',
    'clientes': '/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Clientes.csv'
}

for tabela, caminho in arquivos.items():
    if caminho.endswith('.csv'):
        df = pd.read_csv(caminho)
    elif caminho.endswith('.xlsx'):
        df = pd.read_excel(caminho)
    else:
        print(f'Formato não suportado da tabela {tabela}')
        continue

    #print(df.head)
    print(f"Tabela: {tabela} - Colunas:",df.columns.tolist())


# Função para padronizar os nomes das colunas para minúsculas e com underscores
def padronizar_colunas(df):
    df.columns = [col.lower().replace(' ', '_') for col in df.columns] # Converter todas as colunas para minúsculas e substituir espaços por underscores
    return df

# Função para transformação de Produto
def transformar_produto(df):
    df['price'] = df['price'].astype(float) # Garantir que 'price' seja float
    df['id_category'] = df['id_category'].astype(int) # Garantir que 'id_category' seja inteiro
    df['name'] = df['name'].str.strip().str.title()  # Tornar o 'name' mais uniforme

    # Se caso os id começar com 0 e eles fazer relação com outras tabelas
    df['id'] = df['id'] + 1
    df['id_category'] = df['id_category'] + 1

    return df

# Função para transformação de Ordens
def transformar_ordens(df):
    df['created_at'] = pd.to_datetime(df['created_at'])
    df['customer_id'] = df['customer_id'].astype(int)  # Garantir que 'customer_id' seja int

    # Se caso os id começar com 0 e eles fazer relação com outras tabelas
    df['id'] = df['id'] + 1
    df['customer_id'] = df['customer_id'] + 1
    return df

# Função para transformação de Items
def transformar_items(df):
    df['order_id'] = df['order_id'].astype(int)  # Garantir que 'order_id' seja int
    df['product_id'] = df['product_id'].astype(int)  # Garantir que 'product_id' seja int
    df['quantity'] = df['quantity'].astype(int)  # Garantir que 'quantity' seja int
    df['total_price'] = df['total_price'].astype(float)  # Garantir que 'total_price' seja float
    
    # Se caso os id começar com 0 e eles fazer relação com outras tabelas
    df['id'] = df['id'] + 1
    df['order_id'] = df['order_id'] + 1
    df['product_id'] = df['product_id'] + 1
    return df

# Função para transformação de Categoria
def transformar_categoria(df):
    df['name'] = df['name'].str.strip() # Remove espaços em branco do início e fim de cada string

    # Se caso os id começa com 0
    df['id'] = df['id'] + 1
    return df

# Função para transformação de Clientes
def transformar_clientes(df):
    df['first_name'] = df['first_name'].str.strip().str.title() # Garantir que 'first_name' não tenha espaços e a primeira letra seja maiuscula
    df['last_name'] = df['last_name'].str.strip().str.title() # Garantir que 'last_name' não tenha espaços e a primeira letra seja maiuscula
    df['email'] = df['email'].fillna('Não informado').astype(str).str.lower()  # Padronizar 'email' para minúsculo
    df['email'] = df['email'].replace('não informado', 'Não informado')
    df['number'] = df['number'].fillna(0).astype(float) # Garantir que 'number' seja float e se a coluna for numérica usar 0 para valores ausentes

    # Se caso os id começa com 0
    df['id'] = df['id'] + 1

    # Tratar valores ausentes (NaN) em algumas colunas opcionais
    colunas_opcionais = ['cell_phone', 'country', 'state', 'street', 'additionals']
    for coluna in colunas_opcionais:
        if coluna in df.columns:
            df[coluna] = df[coluna].fillna('Não informado')  # Garante que toda coluna que for NaN seja substituido
    return df

# Função para carregar e enviar com chamada de transformação
def carregar_e_enviar(tabela, caminho, projeto, dataset, credencial, if_exists='replace'):
    print(f'Carregando: {tabela}')
    
    if caminho.endswith('.csv'):
        df = pd.read_csv(caminho, sep=',')  # Define o separador aqui
    elif caminho.endswith('.xlsx'):
        df = pd.read_excel(caminho)
    else:
        raise ValueError(f'Formato não suportado: {caminho}')
    
    df = padronizar_colunas(df)

    # Aplica transformação específica por nome da tabela
    if tabela == 'produto':
        df = transformar_produto(df)
    elif tabela == 'ordens':
        df = transformar_ordens(df)
    elif tabela == 'items':
        df = transformar_items(df)
    elif tabela == 'categoria':
        df = transformar_categoria(df)
    elif tabela == 'clientes':
        df = transformar_clientes(df)

    destino = f'{projeto}.{dataset}.{tabela}'
    print(f'Enviando para BigQuery: {destino}')
    
    df.to_gbq(destination_table=destino,
              project_id=projeto,
              if_exists=if_exists,
              credentials=credencial)

for tabela, caminho in arquivos.items():
    carregar_e_enviar(tabela, caminho, projeto, dataset, credencial)
