==> engenharia-dados/formacao/engenharia-dados-python-banco-dados/Google BigQuery/08.extraindo_varios_arquivos.ipynb <==
from google.cloud import bigquery
import pandas as pd

# Especifique o caminho para o arquivo JSON de credenciais
credencial = "/Users/juniorsilva/Downloads/python-curso-459117-fc7b944c47c6.json"

# Crie um cliente BigQuery usando as credenciais
client = bigquery.Client.from_service_account_json(credencial)

# querys 
queryProduto = """ 
SELECT * FROM `python-curso-459117.Ecommerce.Produto`
LIMIT 10
"""
queryCategoria=""" 
SELECT * FROM `python-curso-459117.Ecommerce.Categoria`
WHERE ID IN (0, 5, 8, 10)
"""
queryItens="""
SELECT * FROM `python-curso-459117.Ecommerce.Items`
"""
queryordens="""
SELECT * FROM `python-curso-459117.Ecommerce.Ordens`
LIMIT 20
"""
queryClientes="""
SELECT * FROM `python-curso-459117.Ecommerce.clientes`
LIMIT 50
"""


resultadoProduto    = client.query(queryProduto)
resultadoCategoria  = client.query(queryCategoria)
resultadoItens      = client.query(queryItens)
resultadoOrdens     = client.query(queryordens)
resultadoClientes   = client.query(queryClientes)

# Converta os resultados em um DataFrame
Produto     = resultadoProduto.to_dataframe()
DfCategoria = resultadoCategoria.to_dataframe()
DfItens     = resultadoItens.to_dataframe()
DfOrdens    = resultadoOrdens.to_dataframe()
DfClientes  = resultadoClientes.to_dataframe()

#DfOrdens.head()

DfOrdens['created_at'] = DfOrdens['created_at'].dt.tz_convert(None)

pasta = "/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Extracao"
ExcelProduto   = 'Produto.xlsx'
ExcelCategoria ='Categoria.xlsx'
ExcelItens     = 'Itens.xlsx'
ExcelOrdens    = 'Ordens.xlsx'
ExcelClientes  = 'Clientes.xlsx'

Produto.to_excel(pasta+ExcelProduto, index=False)  
DfCategoria.to_excel(pasta+ExcelCategoria, index=False)  
DfItens.to_excel(pasta+ExcelItens, index=False)  
DfOrdens.to_excel(pasta+ExcelOrdens, index=False)  
DfClientes.to_excel(pasta+ExcelClientes, index=False)
