==> engenharia-dados/formacao/engenharia-dados-python-banco-dados/Google BigQuery/09.executando_comando_antes_carga.ipynb <==
import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq
from datetime import datetime

df = pd.read_excel("/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Ordens.xlsx")
df.head()

tipos_de_dados = df.dtypes
print(tipos_de_dados)


# filto no meu dataflame antes de realizar a carga (apenas o mês e ano atuais)
"""
hoje = datetime.now()
mes_atual = hoje.month
ano_atual = hoje.year

df_filtrado = df[(df['created_at'].dt.month == mes_atual) & (df['created_at'].dt.year == ano_atual)]
"""

#hoje = datetime.now().year
#print(hoje)

#df_filtrado.head(5)

hoje = datetime.now()
mes_atual = hoje.month
ano_atual = hoje.year

df = df[(df['created_at'].dt.month == mes_atual) & (df['created_at'].dt.year == ano_atual)]

df.head()

# Especifique o caminho para o arquivo JSON de credenciais
credencial = "/Users/juniorsilva/Downloads/python-curso-459117-fc7b944c47c6.json"

# Crie um cliente BigQuery usando as credenciais
clientacao = bigquery.Client.from_service_account_json(credencial)

comandoapagar = """ 

	DELETE
		FROM `python-curso-459117.AulaPython.Ordens`
		WHERE FORMAT_TIMESTAMP('%Y', created_at)  = FORMAT_TIMESTAMP('%Y', CURRENT_DATE())
	AND FORMAT_TIMESTAMP('%m', created_at)  = FORMAT_TIMESTAMP('%m', CURRENT_DATE())

"""
# Execute a consulta
apagardados = clientacao.query(comandoapagar)

https://cloud.google.com/bigquery/docs/reference/standard-sql/format-elements


# credenciais para validar importar dados para o google bigquery
credencial = service_account.Credentials.from_service_account_file(
    "/Users/juniorsilva/Downloads/python-curso-459117-fc7b944c47c6.json", scopes=['https://www.googleapis.com/auth/bigquery'])

# Variaveis do projeto 
projeto = 'python-curso-459117'
dataset = 'AulaPython'
tabela = 'Ordens'
parameretro ='append'      
# replace = subistuir a antiga pela original 
# append = adiciona dados no fim da tabela 

df.to_gbq(destination_table=f'{projeto}.{dataset}.{tabela}',   
          project_id=projeto,
          if_exists=parameretro,
          credentials=credencial)

"""  

SELECT *
	extract(YEAR  FROM created_at) as ano,
	extract(month  FROM created_at) as mes,
	FORMAT_TIMESTAMP('%Y', created_at) AS anoFormat,
	FORMAT_TIMESTAMP('%m', created_at) AS mesformat,
	CURRENT_DATE() hoje
FROM `python-curso-459117.AulaPython.Ordens` 

"""

#Delete
"""

	DELETE FROM `python-curso-459117.AulaPython.Ordens` 
	WHERE FORMAT_TIMESTAMP('%Y', created_at) = FORMAT_TIMESTAMP('%Y', CURRENT_DATE())
	AND FORMAT_TIMESTAMP('%m', created_at) = FORMAT_TIMESTAMP('%m', CURRENT_DATE())

"""
