==> engenharia-dados/formacao/engenharia-dados-python-banco-dados/Google BigQuery/11.consulta_sql_datflame.ipynb <==
import pandas as pd

# Ler o arquivo Excel em um DataFrame
df = pd.read_excel("/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Produto.xlsx")

df.head(15)

#filtrar produto especifico metodo df.query()
dfsapatoroxo= df.query(' Nome == "Sapato Roxo" ')

dfsapatoroxo.head()

#varios produtos em uma lista (nome in ())
df.query("Nome IN ('TV', 'Máquina de Secar', 'Shampoo')" )


df.head(15)

#exemplo de valores
Valor1000= df.query('Price > 1000')

Valor1000.head(15)

df.query('Price < 50')

df.query('Price == 15')

clientes    = pd.read_csv  ("/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Clientes.csv", delimiter=",")
clientes.head(10)

#filtrar ano e mes especifico diferente entre | e &

"""
| Pipe          = ou  ele obeder um criterio ou o outro e traz o resultado se tiver qualqer 1 criterio aceito
& E Comercial = retorna somente quando cumprir todos os criterios 
"""

#varios produtos em uma lista (nome in ())  -- com variaveis de uma lista 
#criar uma lista de filtros ['lista separada por ,'] exemplo de filtrar estados Acre,Bahia,Goiás e depois somenente o primeiro nome @ faz a referencia
filtroestado=['Acre','Bahia','Goiás','Roraima']
clientes.query('state  in @filtroestado   | first_name ==  "Marta"   ')

Datas na Query

import pandas as pd

clientes    = pd.read_csv  ("/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Clientes.csv", delimiter=",")
clientes.head(10)

clientes.dtypes

# Converter a coluna 'data' para o tipo de dados datetime, se ainda não estiver(para funcionar na query tem que estar nesse formato de data)
clientes['created_at'] = pd.to_datetime(clientes['created_at'])

clientes.head()

#clientes['ano']=clientes['created_at'].dt.year

# filtrando 1 ano especifico 
clientes.query('created_at.dt.year  == 2017')

mesdez = clientes.query('created_at.dt.month  == 10')

#exemplo ano 2017 e mes 10
clientes.query('created_at.dt.year == 2017  &  created_at.dt.month  == 10 ')


#exemplo ano 2017 ou mes 10 (simulando obedecendo ou 1 criterio ouo outro) fique atento com e ou ou 
clientes.query('created_at.dt.year == 2017  |  created_at.dt.month  == 10 ')

#### Ano e mes Atual

from datetime import datetime
import pandas  as pd

ordens = pd.read_excel("/Users/juniorsilva/develop/data/estudos-engenharia-dados/engenharia-dados-com-python-e-banco-de-dados/Google BigQuery/Arquivos/Ordens.xlsx")
ordens.head()

# Mês e ano atual
data_atual=datetime.now()
ano_atual = datetime.now().year  # Traz o ano atual
mes_atual = datetime.now().month # Traz o mes atual

# Filtrando mês e ano atual de um DF

ordens.query('created_at.dt.year == @ano_atual   &  created_at.dt.month ==  @mes_atual')
